fix: simplify user in redux to just username
