--- src/actions/user.js
@@ -1,6 +1,0 @@
-import { RECEIVED_USER } from '../reducers/user'
-
-export const acReceivedUser = value => ({
-    type: RECEIVED_USER,
-    value,
-})
