+++ src/actions/username.js
@@ -1,0 +1,6 @@
+import { SET_USERNAME } from '../reducers/username'
+
+export const acSetUsername = value => ({
+    type: SET_USERNAME,
+    value,
+})
